Renaming and add a mdns to pico test
