+++ hw_test/tests/lib_to_porky_usb.rs
@@ -13,7 +13,6 @@ const SERIAL: &str = "e66138528350be2b";
 /// These tests test connecting to USB connected porky devices by USB and TCP, using library
 /// methods to do so.
 ///
-#[cfg(all(feature = "discovery", feature = "usb"))]
 /// Get the IP and Port for a TCP connection to a USB connected porky
 pub async fn get_ip_and_port_by_usb() -> anyhow::Result<Vec<(IpAddr, u16)>> {
     let mut ip_devices: Vec<(IpAddr, u16)> = vec![];
