+++ hw_test/tests/piggui_to_porky_tcp.rs
@@ -7,13 +7,13 @@ use serial_test::serial;
 #[path = "../../piggui/tests/support.rs"]
 mod support;
 
-#[path = "usb.rs"]
-mod usb;
+#[path = "lib_to_porky_usb.rs"]
+mod lib_to_porky_usb;
 
 #[tokio::test]
 #[serial]
 async fn connect_tcp() {
-    let ip_and_ports = usb::get_ip_and_port_by_usb()
+    let ip_and_ports = lib_to_porky_usb::get_ip_and_port_by_usb()
         .await
         .expect("Could not get IP and port of USB connected devices");
